fix fatal account no owner error

+++ api/controllers/AccountController.js
@@ -8,18 +8,18 @@
 module.exports = {
   create: function(req, res) {
     var params = req.params.all();
-    Account.findOne({ name: params.name }).exec(function accountFound(err, account) {
+    Account.findOne({ ownerId: params.ownerId, name: params.name }).exec(function accountFound(err, account) {
       if(err) {
         return res.serverError();
       }
       if (account) {
         return res.badRequest();
       }
-      Account.create(params, function accountCreated(err, account) {
+      Account.create({ ownerId: params.ownerId, name: params.name } , function accountCreated(err, account) {
         if (err) {
           return res.serverError();
         }
-        Account.find().exec(function accountsFound(err, accounts) {
+        Account.find({ ownerId: params.ownerId }).exec(function accountsFound(err, accounts) {
           if (err) {
             return res.serverError();
           }
@@ -30,16 +30,25 @@ module.exports = {
   },
   destroy: function(req, res) {
     var params = req.params.all();
-    Account.destroy({ name: params.name }).exec(function (err) {
+    Account.destroy({ ownerId: params.ownerId, name: params.name }).exec(function (err) {
       if (err) {
         return res.serverError();
       }
-      Account.find().exec(function accountsFound(err, accounts) {
+      Account.find({ ownerId: params.ownerId }).exec(function accountsFound(err, accounts) {
         if (err) {
           return res.serverError();
         }
         return res.ok(accounts);
       });
+    });
+  },
+  findMine: function(req, res) {
+    var params = req.params.all();
+    Account.find({ ownerId: params.ownerId }).exec(function accountsFound(err, accounts) {
+      if (err) {
+        return res.serverError();
+      }
+      return res.ok(accounts);
     });
   },
 };
